+++ lib/src/splay/displaced_marker_offset.dart
@@ -1,7 +1,8 @@
 import 'package:flutter_map/plugin_api.dart';
 import 'package:flutter_map_supercluster/src/splay/displaced_marker.dart';
 
-/// A [Marker] which has been displaced from its original position.
+/// Pixel positions for a [Marker] which has been displaced from its original
+/// position.
 class DisplacedMarkerOffset {
   final DisplacedMarker displacedMarker;
   final CustomPoint displacedOffset;